Add firmware check
Contributes to CL-1222
